district policy

+++ app/controllers/districts_controller.ts
@@ -1,12 +1,20 @@
 import type { HttpContext } from '@adonisjs/core/http'
 import district_service from '#services/district_service'
+import DistrictPolicy from '#policies/district_policy'
+import ForbiddenException from '#exceptions/forbidden_exception'
 
 export default class DistrictsController {
-  async index({}: HttpContext) {
+  async index(ctx: HttpContext) {
+    if (await ctx.bouncer.with(DistrictPolicy).denies('index')) {
+      throw new ForbiddenException()
+    }
     return await district_service.getAllDistricts()
   }
 
   async store(ctx: HttpContext) {
+    if (await ctx.bouncer.with(DistrictPolicy).denies('create')) {
+      throw new ForbiddenException()
+    }
     return await district_service.createDistrict(ctx)
   }
 
@@ -15,10 +23,16 @@ export default class DistrictsController {
   }
 
   async update(ctx: HttpContext) {
+    if (await ctx.bouncer.with(DistrictPolicy).denies('edit')) {
+      throw new ForbiddenException()
+    }
     return await district_service.updateDistrict(ctx)
   }
 
   async destroy(ctx: HttpContext) {
+    if (await ctx.bouncer.with(DistrictPolicy).denies('delete')) {
+      throw new ForbiddenException()
+    }
     return await district_service.deleteDistrict(ctx)
   }
 }
